pyproject.toml
[project]
name = "fsapi"
version = "1.0.0"
description = "Full-stack FastAPI framework with PostgreSQL, Jinja2 templates, and database migrations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.121.3",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-playwright>=0.3.0",
    "ruff>=0.14.6",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
